병 준 용 코드 합침 (네 번째 커밋 및 푸시)

+++ src/main/java/com/example/hong/entity/Restaurant.java
@@ -4,6 +4,8 @@ import com.example.hong.domain.ApprovalStatus;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "restaurants",
         indexes = {
@@ -36,8 +38,8 @@ public class Restaurant {
     private int reviewCount;
 
 
-    private double lat;
-    private double lng;
+    private BigDecimal lat;
+    private BigDecimal lng;
 
 
     @Enumerated(EnumType.STRING)
